BaseTools/tools_def CLANG38: Suppress unaligned access warning
Even though the presence of the 'packed' pragma should be a strong hint
that the misaligned placement of a GUID in a struct is intentional,
recent Clang versions will object nonetheless, and break the build due
to the presence of such GUIDs in the FPDT ACPI tables.

This is obviously not something we can fix in the code, so let's just
suppress the warning/error instead.

Change-Id: Ib1b6fd4b554b852c5db6ed9ad283d2b750748cf4
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>